+++ packages/better-auth/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @vexcms/better-auth
 
+## 0.1.0-alpha.10
+
 ## 0.1.0-alpha.9
 
 ## 0.1.0-alpha.8